Вынес общие функции в utils.php
